requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
extra-streamlit-components==0.1.71
firebase==4.0.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.14.2
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
replicate==1.0.4
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.40.2
streamlit-authenticator==0.4.1
tenacity==9.0.0
tifffile==2024.9.20
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
